requirements.txt
# Flask and web framework
Flask==3.0.0
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
Flask-Session==0.5.0
python-dotenv==1.0.0

# Database
pymongo==4.6.0
redis==5.0.1

# LLM and AI
langchain==0.1.13
langchain-core==0.1.33
langchain-openai==0.1.7
langchain-community==0.1.26
langchain-text-splitters==0.0.1
openai==1.6.1
tavily-python==0.3.10

# Vector store
chromadb==0.4.24

# Security
bcrypt==4.1.1
PyJWT==2.8.1

# Validation
pydantic==2.5.2

# Utilities
requests==2.31.0
tenacity==8.2.3  # Retry logic

# Development and testing
pytest==7.4.4
pytest-cov==4.1.0
pytest-mock==3.13.0

# Production server
gunicorn==21.2.0
